financial-agent-hackathon/requirements.txt
langgraph>=0.2.18
langchain-openai==0.2.0
pandas==2.2.2
openpyxl==3.1.2
streamlit==1.37.1
